Remove resolved_localns, move weak dict handling to NsResolver for pydantic parent ns
